# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/patients": {
      "get": {
        "operationId": "PatientsController_getAllPatients",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all patients"
          }
        },
        "summary": "Get all patients",
        "tags": [
          "Patients"
        ]
      }
    },
    "/patients/{userId}": {
      "get": {
        "operationId": "PatientsController_getPatientProfile",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "Patient user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient profile object"
          }
        },
        "summary": "Get patient profile by user ID",
        "tags": [
          "Patients"
        ]
      },
      "put": {
        "operationId": "PatientsController_updatePatientProfile",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "Patient user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patient profile updated successfully"
          }
        },
        "summary": "Update patient profile",
        "tags": [
          "Patients"
        ]
      }
    },
...
  },
  "info": {
    "title": "MEDBITS Healthcare API",
    "description": "REST API for the MEDBITS Healthcare Management System.\n\n**Authentication**: Pass the user role in the `role` request header.\n\nValid roles: `patient` | `doctor` | `frontdesk` | `admin`",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "User role for RBAC"
      }
    },
    "schemas": {
...
      "UpdatePatientProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "dob": {
            "type": "string",
            "example": "1990-01-01"
          },
          "gender": {
            "type": "string",
            "example": "Male"
          },
          "bloodGroup": {
            "type": "string",
            "example": "O+"
          },
          "phone": {
            "type": "string",
            "example": "9876543210"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "guardianName": {
            "type": "string",
            "example": "Jane Doe"
          }
        }
      },
...
    }
  }
}